DbBase
Untuk CRUD
